+++ backend/src/models/index.ts
@@ -4,7 +4,7 @@ export { Attendance, IAttendance } from './Attendance';
 export { AttendanceFile, IAttendanceFile } from './AttendanceFile';
 export { Leave, ILeave } from './Leave';
 export { PostClock, IPostClock } from './PostClock';
-export { BusinessTrip, IBusinessTrip } from './BusinessTrip';
+export { BusinessTrip, IBusinessTrip, IBusinessTripClockTime } from './BusinessTrip';
 export { OfficialBusiness, IOfficialBusiness } from './OfficialBusiness';
 export { Counter, ICounter } from './Counter';
 export { LeaveAdjustment, ILeaveAdjustment } from './LeaveAdjustment';